requirements.txt
configparser==3.5.0
future==0.16.0
pycrypto==2.6.1
six==1.10.0
